Export creel model estimates

Primary control function of the export, transform, and load (ETL) process. Takes standardized model outputs and user-input parameters to export the model estimates appropriately.

Usage

export_estimates(params, analysis_lut, creel_estimates, conn = NULL)

Arguments

params

User-input parameters defined during the model estimation process.

analysis_lut

Lookup table created during the model estimation process which stores a session-specific analysis_id key and metadata about the analysis.

creel_estimates

List object containing model estimates in a standardized format. Typically passed from 'transform_estimates' function.

conn

Database connection object. If NULL (default), a new connection will be established.

Value

??

See also